+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.4.1...HEAD)
 ### Changed
 * [STREAM-997](https://inindca.atlassian.net/browse/STREAM-997) - Updated requestWithApiRetry to retry on network error codes. Added maxAttempts option to requestApiWithRetry.
+* [STREAM-1006](https://inindca.atlassian.net/browse/STREAM-1006) - Added ECONNRESET to the list of network error codes to retry on.
 
 # [v19.4.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.4.0...v19.4.1)
 ### Changed