+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.4.1...HEAD)
+# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.5.0...HEAD)
+
+# [v19.5.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.4.1...v19.5.0)
 ### Added
-* [STREAM-995](https://inindca.atlassian.net/browse/STREAM-995) - Include `missingPermissions` on topic subscribe Error
+* [STREAM-995](https://inindca.atlassian.net/browse/STREAM-995) - Include `missingPermissions` on topic subscribe Error.
+
 ### Changed
 * [STREAM-997](https://inindca.atlassian.net/browse/STREAM-997) - Updated requestWithApiRetry to retry on network error codes. Added maxAttempts option to requestApiWithRetry.
 * [STREAM-1006](https://inindca.atlassian.net/browse/STREAM-1006) - Added ECONNRESET to the list of network error codes to retry on.
+
 ### Fixed
 * [STREAM-1026](https://inindca.atlassian.net/browse/STREAM-1026) - Filter out all IP addresses (local, remote, and related) from stats event details before sending to New Relic. See [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/API/RTCIceCandidatePairStats) for more details on potential IP address exposure.